fixed issue that would cause thrown objects to spawn in the wrong place
the objects' position was not updating to reflect the player's world
coordinates after the screen had scrolled. the fix was to use the world
coordinates of the Character rather than the AnimateEntity.

also fixed thrown objects being deleted for being off screen when they
should not have been (due to scrolling)